[TEST] Move dynamic templates tests to DynamicTemplatesTests
Some tests around dynamic templates used to be in RootObjectMapperTests while others are in DynamicTemplatesTests. This commits moves all tests to be under DynamicTemplatesTests.
